Split shell shared key limit tests
